refactor iconForDomain into domainhelper.tsx
